make BaseRecordReader its own class
